Make sure a pointer fits in the argument!

+++ MainPanel.c
@@ -162,7 +162,7 @@ const char* MainPanel_getValue(MainPanel* this, int i) {
    return "";
 }
 
-bool MainPanel_foreachProcess(MainPanel* this, MainPanel_ForeachProcessFn fn, int arg, bool* wasAnyTagged) {
+bool MainPanel_foreachProcess(MainPanel* this, MainPanel_ForeachProcessFn fn, size_t arg, bool* wasAnyTagged) {
    Panel* super = (Panel*) this;
    bool ok = true;
    bool anyTagged = false;
