+++ MainPanel.h
@@ -34,7 +34,7 @@ int MainPanel_selectedPid(MainPanel* this);
 
 const char* MainPanel_getValue(MainPanel* this, int i);
 
-bool MainPanel_foreachProcess(MainPanel* this, MainPanel_ForeachProcessFn fn, int arg, bool* wasAnyTagged);
+bool MainPanel_foreachProcess(MainPanel* this, MainPanel_ForeachProcessFn fn, size_t arg, bool* wasAnyTagged);
 
 extern PanelClass MainPanel_class;
